Fix : refactoring reviewList state
