+++ frontend/src/pages/placeDetailPage/placeDetailPage.jsx
@@ -106,25 +106,6 @@ const PlaceDetailPage = (props) => {
     window.scrollTo(0, 0);
   }, []);
 
-  // const [reviewDatas, setReviewDatas] = useState([
-  //   {
-  //     createdAt: new Date(2011, 0, 1, 0, 0, 0, 0),
-  //     likeCount: 5,
-  //   },
-  //   {
-  //     createdAt: new Date(2011, 0, 1, 0, 0, 0, 2),
-  //     likeCount: 6,
-  //   },
-  //   {
-  //     createdAt: new Date(2011, 0, 1, 0, 0, 0, 1),
-  //     likeCount: 7,
-  //   },
-  //   {
-  //     createdAt: new Date(2011, 0, 1, 0, 0, 0, 4),
-  //     likeCount: 3,
-  //   },
-  // ]);
-
   return (
     <div className={styles.PlaceDetailPage}>
       <DetailHeader
